requirements.txt
psycopg2-binary>=2.9
python-dateutil>=2.8
pyyaml>=6.0
requests>=2.31
